requirements.txt
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-JWT-Extended==4.6.0
flask-swagger-ui==4.11.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
PyJWT==2.9.0
requests==2.32.3
soupsieve==2.6
urllib3==2.2.3
Werkzeug==3.0.6
gunicorn==20.1.0
flask_cors==5.0.1
python-dotenv==1.0.0
pytest==7.4.0
pytest-cov==4.1.0
requests-mock==1.11.0